requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.115.12
google==3.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
limits==5.2.0
packaging==25.0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2